Created a base stats class, that contains all the stats of an entity
